requirements.txt
flask
requests
gunicorn
ccxt>=4.5.0
python-dotenv>=1.0.0
